+++ base/Splendor_v2/env.py
@@ -85,7 +85,7 @@ def get_player_state(env_state, lv1, lv2, lv3):
     p_state = np.append(p_state, other_scores) #Lấy điểm của người chơi khác
     p_state = np.append(p_state, (env_state[161:164] != 100)*1) #Lấy thông tin của các thẻ ẩn có thẻ úp, nếu có thể úp thì là 1
     p_state = np.append(p_state, len(np.where(env_state[:90] == 5)[0])) #Số lượng thẻ có thể úp trong bàn
-    return p_state
+    return p_state.astype(np.float64)
 
 @njit()
 def check_victory(p_state):
